api/requirements.txt
requests
boto3
# boto3~=1.28.21
# botocore~=1.31.21
python-dotenv
python-multipart
uvicorn
fastapi~=0.97.0
sqlalchemy~=2.0.16
PyPDF2
boto3
beautifulsoup4
pydantic
pdfrw